requirements.txt
Django==1.8.0
django-model-utils>=2.2
django-extensions==1.5.2
psycopg2==2.5.4
Markdown==2.5.2
python-dateutil==2.4.0
djangorestframework>=3.1.1
django-filter==0.9.2
django-braces==1.4.0
requests==2.5.1
requests-cache==0.4.9
lxml==3.4.2
boltons==0.6.4
gunicorn==18.0
supervisor==3.0
gevent==0.13.8
greenlet==0.4.1
django-compressor==1.5
python-memcached==1.54
Unipath==1.1
psycopg2==2.6.1
